+++ src/openmc_fusion_benchmarks/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Backend-specific interfaces (OpenMC, Serpent, MCNP, ...)."""
