+++ src/openmc_fusion_benchmarks/backends/openmc/__init__.py
@@ -1,0 +1,3 @@
+"""OpenMC backend adapters for openmc_fusion_benchmarks."""
+
+from .tallies import *
